+++ uptimerobot/api/alert_contact.go
@@ -18,9 +18,11 @@ var alertContactType = map[string]int{
 	"webhook":    5,
 	"pushbullet": 6,
 	"zapier":     7,
-	"pushover":   8,
+	"pushover":   9,
 	"hipchat":    10,
 	"slack":      11,
+	"telegram":		18,
+	"hangouts":	  21,
 }
 var AlertContactType = mapKeys(alertContactType)