+++ uptimerobot/resource_uptimerobot_status_page.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/hashicorp/terraform/helper/validation"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/louy/terraform-provider-uptimerobot/uptimerobot/api"
 )
 
@@ -31,8 +31,9 @@ func resourceStatusPage() *schema.Resource {
 				Default:  "",
 			},
 			"password": &schema.Schema{
-				Type:     schema.TypeString,
-				Optional: true,
+				Type:      schema.TypeString,
+				Optional:  true,
+				Sensitive: true,
 			},
 			"sort": {
 				Type:         schema.TypeString,